refactor(napari): move layer sync callbacks to movement/napari/layer_wiring.py
Pure move: the eight layer-sync handlers become module-level functions
(the three viewer-dependent ones take `viewer` as first argument), and
the movement layer metadata keys move with them. `DataLoader` keeps thin
methods delegating to the new functions, so behaviour and the widget's
API are unchanged.
