+++ movement/napari/loader_widgets.py
@@ -1,14 +1,11 @@
 """Widgets for loading movement datasets from file."""
 
-import warnings
 from pathlib import Path
 
 import numpy as np
 import pandas as pd
 import xarray as xr
-from napari.components.dims import RangeTuple
-from napari.layers import Image, Points, Shapes, Tracks
-from napari.layers.base import ActionType
+from napari.layers import Points
 from napari.settings import get_settings
 from napari.utils.notifications import show_error, show_warning
 from napari.viewer import Viewer
@@ -26,10 +23,23 @@ from qtpy.QtWidgets import (
 from movement.io.load import load_dataset, rename_legacy_dimensions
 from movement.napari.convert import ds_to_napari_layers
 from movement.napari.layer_styles import (
-    EDITED_POINT_SYMBOL,
     BoxesStyle,
     PointsStyle,
     TracksStyle,
+)
+from movement.napari.layer_wiring import (
+    DATASET_ATTRS_KEY,
+    POINTS_LAYER_KEY,
+    POINTS_PROPERTIES_KEY,
+    TRACKS_LAYER_KEY,
+    frame_axis_is_sliced,
+    on_points_data_changed,
+    remove_from_tracks_layer,
+    set_point_symbol_by_edited,
+    set_tracks_layer_data,
+    sync_tracks_layer,
+    update_frame_slider_range,
+    update_points_layers_editable,
 )
 from movement.utils.logging import logger
 from movement.validators.datasets import ValidBboxesInputs, ValidPosesInputs
@@ -54,17 +64,6 @@ SUPPORTED_DATA_FILES = {
     **SUPPORTED_BBOXES_FILES,
     **SUPPORTED_NETCDF_FILES,
 }
-
-# Metadata keys stored on the movement Points layer.
-# - POINTS_LAYER_KEY marks the layer as movement-created.
-# - POINTS_PROPERTIES_KEY holds the full properties df, incl. the NaN rows
-#   dropped from the live layer, needed to reconstruct the dataset.
-# - DATASET_ATTRS_KEY holds the source dataset's attrs (source_software, fps…).
-# - TRACKS_LAYER_KEY holds a reference to the companion Tracks layer.
-POINTS_LAYER_KEY: str = "movement_points_layer"
-POINTS_PROPERTIES_KEY: str = "movement_points_properties"
-DATASET_ATTRS_KEY: str = "movement_dataset_attrs"
-TRACKS_LAYER_KEY: str = "movement_tracks_layer"
 
 
 class DataLoader(QWidget):
@@ -392,147 +391,34 @@ class DataLoader(QWidget):
         logger.info("Added tracked dataset as a napari Points layer.")
 
     def _frame_axis_is_sliced(self) -> bool:
-        """Whether frame is the sliced axis in a 2D view.
-
-        A point drag only ever touches the currently *displayed* axes
-        (see ``Points._move`` in napari). When frame is the sliced
-        axis, that means x/y -- everything is safe to edit. If axes
-        have been rolled so frame is displayed instead, or a 3D view
-        is active, a drag could move a point onto a different frame.
-        """
-        return (
-            self.viewer.dims.ndisplay == 2 and self.viewer.dims.order[0] == 0
-        )
+        """Whether frame is the sliced axis in a 2D view."""
+        return frame_axis_is_sliced(self.viewer)
 
     def _update_points_layers_editable(self, event=None):
-        """Disable point editing while the frame axis isn't sliced.
-
-        Connected to ``viewer.dims.events.order``/``ndisplay``.
-        In the default view, the frame axis is the slider, so
-        dragging a point can only change its x/y position. Rolling
-        the axes or switching to 3D makes frame draggable too, which
-        would let a drag move a point to another frame. Disable
-        editing on every movement Points layer while that is the
-        case; napari greys out the select/add/delete controls.
-        """
-        frame_axis_is_sliced = self._frame_axis_is_sliced()
-        for layer in self.viewer.layers:
-            if isinstance(layer, Points) and layer.metadata.get(
-                POINTS_LAYER_KEY
-            ):
-                layer.editable = frame_axis_is_sliced
+        """Disable point editing while the frame axis isn't sliced."""
+        update_points_layers_editable(self.viewer, event)
 
     @staticmethod
     def _set_point_symbol_by_edited(layer: Points) -> None:
         """Show points flagged as edited with a distinct marker symbol."""
-        edited = layer.properties.get("edited")
-        if edited is None or not edited.any():
-            return
-        symbols = np.asarray(layer.symbol).copy()
-        symbols[edited] = EDITED_POINT_SYMBOL
-        layer.symbol = symbols
+        set_point_symbol_by_edited(layer)
 
     def _on_points_data_changed(self, event):
-        """Keep the corresponding Tracks layer in sync with the Points layer.
-
-        Connected to ``points_layer.events.data``. Handles two actions:
-
-        - ``ActionType.CHANGED`` (a point was dragged): sets the
-          confidence score of moved points to NaN, marks them as
-          edited, and changes their marker symbol to
-          ``EDITED_POINT_SYMBOL`` so edited points are visually
-          distinguishable. The Tracks layer row is updated in place
-          via `_sync_tracks_layer`.
-        - ``ActionType.REMOVED`` (one or more points were deleted):
-          removes the corresponding rows from
-          the Tracks layer via `_remove_from_tracks_layer`.
-        """
-        layer = event.source
-        if not isinstance(layer, Points):
-            return
-
-        if event.action == ActionType.CHANGED:
-            moved_indices = list(event.data_indices)
-            props = layer.properties
-            props["confidence"] = props["confidence"].copy()
-            props["confidence"][moved_indices] = float("nan")
-            if "edited" in props:
-                props["edited"] = props["edited"].copy()
-            else:
-                props["edited"] = np.full(len(props["confidence"]), False)
-            props["edited"][moved_indices] = True
-            layer.properties = props
-            self._set_point_symbol_by_edited(layer)
-            self._sync_tracks_layer(layer, moved_indices)
-
-        elif event.action == ActionType.REMOVED:
-            removed_indices = list(event.data_indices)
-            self._remove_from_tracks_layer(layer, removed_indices)
+        """Keep the Tracks layer in sync with the Points layer."""
+        on_points_data_changed(event)
 
     def _sync_tracks_layer(self, points_layer, moved_indices):
-        """Update the corresponding Tracks layer to match an edited point.
-
-        A moved point's new (frame, y, x) is written to the same row
-        in the Tracks layer, so the track segment connecting the
-        previous frame to this one terminates at the dragged position.
-        """
-        tracks_layer = points_layer.metadata[TRACKS_LAYER_KEY]
-
-        # Points and Tracks layers are built from the same NaN-filtered
-        # array in the same row order (see _add_points_layer/
-        # _add_tracks_layer). The Tracks layer only has an extra
-        # leading track_id column.
-        tracks_data = tracks_layer.data
-        tracks_data[moved_indices, 1:] = points_layer.data[moved_indices]
-
-        self._set_tracks_layer_data(
-            tracks_layer, tracks_data, tracks_layer.properties
-        )
+        """Update the Tracks layer to match an edited point."""
+        sync_tracks_layer(points_layer, moved_indices)
 
     def _remove_from_tracks_layer(self, points_layer, removed_indices):
-        """Remove the rows corresponding to deleted points.
-
-        Users edit the Points layer directly, either dragging points
-        or removing inaccurate predictions. The Tracks layer has no
-        interactive editing of its own, so it must be kept in sync
-        with the Points layer instead.
-
-        ``removed_indices`` are indices in the Points layer which line
-        up with rows in the Tracks layer, the same way
-        :meth:`_sync_tracks_layer` relies on for edits.
-        """
-        tracks_layer = points_layer.metadata[TRACKS_LAYER_KEY]
-
-        tracks_data = np.delete(tracks_layer.data, removed_indices, axis=0)
-        tracks_properties = {
-            key: np.delete(np.asarray(values), removed_indices, axis=0)
-            for key, values in tracks_layer.properties.items()
-        }
-
-        self._set_tracks_layer_data(
-            tracks_layer, tracks_data, tracks_properties
-        )
+        """Remove the rows corresponding to deleted points."""
+        remove_from_tracks_layer(points_layer, removed_indices)
 
     @staticmethod
     def _set_tracks_layer_data(tracks_layer, data, properties):
-        """Set a Tracks layer's data and properties, preserving color_by.
-
-        Setting ``.data`` on a napari Tracks layer resets its internal
-        features to empty, which transiently invalidates ``color_by``
-        (napari warns and falls back to "track_id") even though we
-        restore the same properties and colour-by property right
-        after. Suppress that spurious warning around the sequence.
-        """
-        color_by = tracks_layer.color_by
-        with warnings.catch_warnings():
-            warnings.filterwarnings(
-                "ignore",
-                message=".*Previous color_by key.*",
-                category=UserWarning,
-            )
-            tracks_layer.data = data
-            tracks_layer.properties = properties
-            tracks_layer.color_by = color_by
+        """Set a Tracks layer's data and properties, preserving color_by."""
+        set_tracks_layer_data(tracks_layer, data, properties)
 
     def _add_tracks_layer(self):
         """Add the tracked data to the viewer as a Tracks layer."""
@@ -581,53 +467,8 @@ class DataLoader(QWidget):
         logger.info("Added tracked dataset as a napari Shapes layer.")
 
     def _update_frame_slider_range(self):
-        """Check the frame slider range and update it if necessary.
-
-        This is required because if the data loaded starts or ends
-        with all NaN values, the frame slider range will not reflect
-        the full range of frames.
-        """
-
-        def _layer_has_data(layer):
-            if isinstance(layer, Shapes):
-                return len(layer.data) > 0
-            return layer.data.shape[0] > 0
-
-        # Only update the frame slider range if there are layers
-        # that are Points, Tracks, Image or Shapes with data
-        list_layers = [
-            ly
-            for ly in self.viewer.layers
-            if isinstance(ly, Points | Tracks | Image | Shapes)
-            and _layer_has_data(ly)
-        ]
-        if len(list_layers) > 0:
-            # Get the maximum frame index from all candidate layers
-            max_frame_idx = max(
-                # For every layer, get max_frame_idx metadata if it exists,
-                # else deduce it from the data shape
-                [
-                    getattr(ly, "metadata", {}).get(
-                        "max_frame_idx", ly.data.shape[0] - 1
-                    )
-                    if not isinstance(ly, Shapes)
-                    # Napari stores shapes layer data as a list of 2D arrays
-                    # instead of a 3D array, so we can't use data.shape here
-                    else getattr(ly, "metadata", {}).get(
-                        "max_frame_idx", len(ly.data) - 1
-                    )
-                    for ly in list_layers
-                ]
-            )
-
-            # If the frame slider range is not set to the full range of frames,
-            # update it.
-            if (self.viewer.dims.range[0].stop != max_frame_idx) or (
-                int(self.viewer.dims.range[0].start) != 0
-            ):
-                self.viewer.dims.range = (
-                    RangeTuple(start=0.0, stop=max_frame_idx, step=1.0),
-                ) + self.viewer.dims.range[1:]
+        """Check the frame slider range and update it if necessary."""
+        update_frame_slider_range(self.viewer)
 
     @staticmethod
     def _enable_layer_tooltips():
